[TAY-31]: Backtracking on TAY-30 - removing selection manager from DatasourceType.
I really like having the isSelected boolean in the cell index type. However, it comes at too high cost, as really the selection manager needs to be generic over ItemType to be able track selections across datasources. This means that the SelectionManager must be generic, and that it's item be Hashable, which puts an overly restrictive contraint on the datasource types for Swift 1.2.

In Swift 2.0, we can possibly add the selection manager to the datasource types in a protocol extension.